Don't need to set GOPATH
Base system doesn't have golang installed. This means that only devbox
projects that specifically use golang won't need to worry about the base
system stomping over the env-vars.

The downside is the fact that the base system doesn't have go installed
and has to rely on the use of devbox or something similar. I'm not sure
I like this, but can live with it for now.